+++ packages/common/README.md
@@ -2,6 +2,10 @@
 
 > TODO: description
 
+## Links
+
+[audius-query docs](./src/audius-query/README.md)
+
 ## Usage
 
 ```
