+++ packages/common/src/api/collection.ts
@@ -1,6 +1,5 @@
 import { Kind } from 'models'
-
-import { createApi } from './createApi'
+import { createApi } from 'src/audius-query/createApi'
 
 const collectionApi = createApi({
   reducerPath: 'collectionApi',
