+++ packages/common/src/api/index.ts
@@ -1,6 +1,4 @@
-export * from './AudiusQueryContext'
 export * from './relatedArtists'
 export * from './track'
 export * from './collection'
 export * from './user'
-export * from './hooks'
